+++ react/__mocks__/categoriesTree.js
@@ -66,3 +66,5 @@ export default [
     ],
   },
 ]
+
+export const numberOfFilters = 10