+++ react/components/SearchResult.js
@@ -122,6 +122,7 @@ class SearchResult extends Component {
       runtime: { hints: { mobile } },
       gap,
       quantityOfItemsPerRow,
+      showTitle,
     } = this.props
     const {
       mobileLayoutMode,
@@ -152,12 +153,18 @@ class SearchResult extends Component {
     const hideFacets = !map || !map.length
     const showLoading = loading && !fetchMoreLoading
     const showContentLoader = showLoading && !showLoadingAsOverlay
+
     return (
       <LoadingOverlay loading={showLoading && showLoadingAsOverlay}>
         <div className={`${searchResult.container} w-100 mw9`}>
           <div className={`${searchResult.breadcrumb} db-ns dn-s`}>
             <ExtensionPoint id="breadcrumb" {...breadcrumbsProps} />
           </div>
+          <ExtensionPoint
+            id="search-title"
+            params={params}
+            showTitle={showTitle}
+          />
           <ExtensionPoint id="total-products"
             recordsFiltered={recordsFiltered}
           />